requirements.txt
GitPython==3.1.30
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-pygments==0.2.2
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.26.0
opencv-contrib-python==4.8.0.76
opencv-python==4.7.0.68
openpyxl==3.0.10
pandas==1.5.0
pickleshare==0.7.5
Pillow==9.2.0
scikit-learn==1.2.0
scipy==1.11.3
seaborn==0.12.2
timm==1.0.12
torch==1.13.1
torcheval==0.0.7
torchvision==0.14.1
tqdm==4.64.1
xgboost==1.7.5